refactor: split basis models and database interface in seperate files
